add build scripts for libnl
libnl may not be the first package that you figured we would
build.  But it builds quickly.  Which makes it a suitable
candidate for developing the per-package build system.

Also, not particularly related to this patchset, make it so we
don't have to be root to build stuff.  Just warn the user that we
are about to sudo.