make installer leave the host number in the provisioned file system
We expect that lots of people are going to have lots of mesh
nodes around.  Each one is going to be provisioned, possibly by
the installer rc.local script.  It is handy to assign a different
number to each node so IP addressses and hostnames can be
generated in a coherent way.