introduce meshkit package
This is the core of the zotac mesh nodes (and possibly others in
the future).  It will ultimately implement the bridging of the AP
and mesh interfaces, and also a control interface for mesh
development.  But for now we just document it and decide what the
config variables will look like.

Also, deploy meshkit and a suitable configuration for zotac.