manage static IP address based on hostnumber
Again: this is so we can simply rsync the rootfs over to the
devices without overwriting the /etc/network/interfaces file.  A
byproduct of this is that provisioining is much simpler.

Note that I re-added the -n option to the provisioining script.
But now it is the host number, not the host name.