kernel: add module offsets in .gdbinit
This will allow you to easily decode a backtrace without
manually loading the module and offset each time. The
module offsets may change, so in the future let us
establish a common point to develop against (say
open80211s/master or "linux-v3.x" and adjust these offsets
to match that.