qemu: enable certain virtfs mounts
A qemu target will now attempt to mount two virtfs
targets: 'modules' and 'misc'. 'modules' is simply
$STAGING/lib/modules, while 'misc' defaults to /home and
can be configured in distro11s.conf.

This means a module rebuild (as long as the kernel HEAD
commit SHA is the same) does not require a rootfs rebuild
or running ./scripts/install-modules-qemu.sh.

Also, the 'misc' path can now be used instead of the sshfs
mount to access your hwsim_tests. This deprecates the
sshfs mount complexity, which can now be removed.

To use this feature however, your qemu must be compiled
with virtio support. See
https://github.com/cozybit/open80211s/wiki/QemuAsDevelTarget#virtfs-as-root
for instructions.

Thanks to Bob Copeland for actually developing this
solution.