distro11s: allow FORCE_BUILD of fetch.sh
For jenkins specifically (and perhaps other users sometimes) it is
important to be able to reset the source dirs specified in pkglist to
the remote url, branch and tag specified in the pkglist.

This patch allows this with the following invocation:
FORCE_BUILD=1 ./scripts/fetch.sh

WARNING: Using FORCE_BUILD with fetch with local changes risks loosing
work.  Please ensure all commits from SRCDIR are pushed elsewhere before
using FORCE_BUILD=1 with fetch.

Signed-off-by: Jason Abele <jason@cozybit.com>